created function source_aggregator
First, 2 subsets are created: those with more than 1 centroid of the source data within the boundaries of the lots, and those with 1 or 0.

Then, the ones with morethan 1 go through the function that will aggregate the info of the blocks contained in the lot. For this, blocks are subset using the boolean list that .intersects() generates.